chore: rename Grants.ts file to Project.ts
