+++ src/export-kpis.ts
@@ -8,7 +8,7 @@ import { BalanceParsed } from './export-balances'
 import { TransactionParsed } from './export-transactions'
 import { VotesParsed } from './export-votes'
 import { GovernanceProposalType, Status } from './interfaces/GovernanceProposal'
-import { Project } from './interfaces/Grant'
+import { Project } from './interfaces/Project'
 import { KPI } from './interfaces/KPIs'
 import { MemberInfo, MemberVP } from './interfaces/Members'
 import { ProposalParsed } from './interfaces/Proposal'
