+++ src/export-transactions.ts
@@ -6,7 +6,7 @@ import { Tags, TagType } from './entities/Tags'
 import { Tokens, TokenSymbols } from './entities/Tokens'
 import { Wallets } from './entities/Wallets'
 import { GovernanceProposalType, Status } from './interfaces/GovernanceProposal'
-import { Project } from './interfaces/Grant'
+import { Project } from './interfaces/Project'
 import { EventItem } from './interfaces/Transactions/Events'
 import { TransactionItem } from './interfaces/Transactions/Transactions'
 import { TransferItem, TransferType } from './interfaces/Transactions/Transfers'
